refined bids conversion functions, added template

+++ functions/bidsconversion/bidsconvert_writesessionfiles.m
@@ -3,12 +3,6 @@ function bidsconvert_writesessionfiles(dataReadDir, dataWriteDir, T1WriteDir, su
 
 %% Create SESSION-SPECIFIC files %%%%%%%%%%%%%%%%%%
 
-% Check whether we have write directories
-if ~exist(dataWriteDir, 'dir'); mkdir(dataWriteDir);end
-if ~exist(T1WriteDir, 'dir'); mkdir(T1WriteDir);end
-
-%% Create DATASET-SPECIFIC files %%%%%%%%%%%%%%%%%%
-
 %   - T1w.nii.gz
 %   - coordsystem.json
 %   - electrodes.tsv
@@ -17,15 +11,19 @@ if ~exist(T1WriteDir, 'dir'); mkdir(T1WriteDir);end
 
 % Locate T1 provided by SoM
 T1File = dir(fullfile(dataReadDir, 'T1.nii.gz'));
-if length(T1File) > 1, disp('Warning: multiple T1s found: using first one'); end
+if length(T1File) > 1, fprintf('[%s] Warning: multiple T1s found: using first one\n', mfilename); end
 
 % Copy file: rename and save in appropriate folder
 if length(T1File) < 1
     disp('Warning: no T1 found!') 
 else
     T1_name = fullfile(T1WriteDir, sprintf('sub-%s_ses-%s_T1w.nii.gz', sub_label, ses_labelt1));
-    disp(['Writing ' T1_name '...']);
-    copyfile(fullfile(dataReadDir, 'T1.nii.gz'), T1_name);
+    if exist(T1_name, 'file')
+        fprintf('[%s] T1 already exists in %s, not overwriting.\n', mfilename, T1_name);
+    else
+        fprintf('[%s] Writing %s \n', mfilename, T1_name);
+        copyfile(fullfile(dataReadDir, 'T1.nii.gz'), T1_name);
+    end
 end
 
 % Note on pial file: surface recons should go in derivatives folder (see
@@ -48,7 +46,7 @@ coordsystem_json.IntendedFor = fullfile(sprintf('sub-%s', sub_label), sprintf('s
     sprintf('sub-%s_ses-%s_T1w.nii.gz', sub_label, ses_labelt1)); % this path must be specified relative to the project folder
 
 % Write coordsystem.json file
-disp(['Writing ' coord_json_name '...']);
+fprintf('[%s] Writing %s\n', mfilename, coord_json_name);
 jsonwrite(coord_json_name,coordsystem_json,json_options);
 
 % Note on coordystem json file: SOM also provides electrode locations in
@@ -61,22 +59,11 @@ jsonwrite(coord_json_name,coordsystem_json,json_options);
 
 %% Create electrodes.tsv file
 
-% % Generate table with SOM defaults
-% n = length(elecInx);
-% [electrode_table] = createBIDS_ieeg_electrodes_tsv_nyuSOM(n);
-% 
-% % Fill with relevant information:
-% electrode_table.name = elec_labels(elecInx);
-% electrode_table.x = elec_xyz(elecInx,1);
-% electrode_table.y = elec_xyz(elecInx,2);
-% electrode_table.z = elec_xyz(elecInx,3);
-% electrode_table.type = elec_types(elecInx);
-
 % Generate output name
 electrodes_tsv_name = fullfile(dataWriteDir, sprintf('sub-%s_ses-%s_electrodes.tsv', sub_label, ses_label));
 
 % Write tsv file
-disp(['Writing ' electrodes_tsv_name '...']);
+fprintf('[%s] Writing %s\n', mfilename, electrodes_tsv_name);
 writetable(electrode_table,electrodes_tsv_name,'FileType','text','Delimiter','\t');
 
 end
